requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==2.2
django-filebrowser-no-grappelli==3.7.4
django-tinymce4-lite==1.7.0
djangorestframework==3.11.0
geoip2==3.0.0
idna==2.9
jsmin==2.2.2
maxminddb==1.5.2
mysqlclient==1.4.6
Pillow==7.1.1
pkg-resources==0.0.0
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.9
